Allow the usage of 'ratio' in transforms.
If you specify a 'ratio' parameter in transforms, the plugin will
now calculate the correct height or width based on said ratio.

If both width and height is specified, width is preferred and height
is calculated based on the width and ratio.

Currently this will also set 'fit' mode to 'crop'